+ Dividir o código - pacote Parser e Model
