+++ src/main/kotlin/parser/CurriculoParser.kt
@@ -1,0 +1,5 @@
+package parser
+
+class CurriculoParser {
+
+}
